+++ examples/morpheusvm/throughput/helper.go
@@ -11,8 +11,8 @@ import (
 	mauth "github.com/ava-labs/hypersdk/examples/morpheusvm/auth"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/consts"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/vm"
-	"github.com/ava-labs/hypersdk/loadgen"
 	"github.com/ava-labs/hypersdk/pubsub"
+	"github.com/ava-labs/hypersdk/throughput"
 	"github.com/ava-labs/hypersdk/utils"
 )
 
@@ -22,7 +22,7 @@ type SpamHelper struct {
 	ws      *ws.WebSocketClient
 }
 
-var _ loadgen.SpamHelper = &SpamHelper{}
+var _ throughput.SpamHelper = &SpamHelper{}
 
 
 func (sh *SpamHelper) CreateAccount() (*auth.PrivateKey, error) {
